rename behav. component
