+++ v2/client/src/components/common/BehavioralInsight/index.js
